+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v1.31.0-beta.1 - 2026-08-26
+
+### 🚀 Features
+
+- **release**: Betas gehen im Community-Finale auf (3a686e2f)
+
+### 🔧 Sonstiges
+
+- **docker**: READMEs für die Docker-Hub-Repos, HTTPS-Korrektur (d9b32f6b)
+- v1.30.8 - Version & Changelog vorbereitet (0120aa3f)
+
 ## v1.30.8 - 2026-08-26
 
 ### 🔧 Sonstiges